chore: switch local emcc or docker emcc

+++ wasm-libicu/container.sh
@@ -2,8 +2,13 @@
 
 sh -x prepare.sh $BRANCH
 
-docker rmi trzeci/emscripten
-docker run --rm -v `pwd`:/src trzeci/emscripten \
-  bash -c "(cd icu/icu4c/source && bash -x ../../../build.sh)"
+buildCmd="cd icu/icu4c/source && bash -x ../../../build.sh"
 
-#cd icu/icu4c/source && bash -x ../../../build.sh
+if `which emcc`
+then
+  docker rmi trzeci/emscripten
+  docker run --rm -v `pwd`:/src trzeci/emscripten \
+    bash -c "$buildCmd"
+else
+    bash -c "$buildCmd"
+fi
